+++ src/qt/view/QtIcons.h
@@ -115,6 +115,18 @@ inline QIcon trends() {
             .arg(navColor()));
 }
 
+inline QIcon history() {
+    return fromSvg(
+        QStringLiteral(
+            "<svg xmlns='http://www.w3.org/2000/svg' viewBox='0 0 24 24'"
+            " fill='none' stroke='%1' stroke-width='2' stroke-linecap='round'"
+            " stroke-linejoin='round'>"
+            "<path d='M3 3v5h5'/>"
+            "<path d='M3.05 13A9 9 0 1 0 6 5.3L3 8'/>"
+            "<polyline points='12 7 12 12 15 14'/></svg>")
+            .arg(navColor()));
+}
+
 inline QIcon settings() {
     return fromSvg(
         QStringLiteral(